add new TPM event type constants
Couldn't find these in the current specification, but apparently they're
already in use by newer firmwares, and I was able to find all these
constants in 'canonical/tcglog-parser'.

In the future, it might be better to just quietly ignore unknown event
types, as generally no special handling is needed for them.

Ref: #6